requirements.txt
# Tool Calling Bot Dependencies

# LLM API clients
openai>=1.0.0

# HTTP requests for web search
requests>=2.25.0

# Date and time handling
python-dateutil>=2.8.0
pytz>=2021.1

# Enhanced calculator functions
sympy>=1.9  

# Environment variable management
python-dotenv>=0.19.0 